Adding asset to envelope now functional
